Move sieves from prime.h to sieve.h
Many functions (e.g., is_prime and factor) will have a normal version and a
sieve accelerated version. To avoid mutual dependencies between header
files, we will put the sieve accelerated versions and the sieve classes in
sieve.h.
