+++ number_theory/CMakeLists.txt
@@ -3,6 +3,7 @@ set(SOURCES
   numeric_unittest.cpp
   modular_unittest.cpp
   prime_unittest.cpp
+  sieve_unittest.cpp
   # ... more
 )
 add_unittest(${SOURCES})
